requirements.txt
# Gmail Email Agent Requirements
# Core dependencies
google-api-python-client==2.110.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.25.2

# AI and ML
openai==1.6.1
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2

# Email processing
email-validator==2.1.0
beautifulsoup4==4.12.2
html2text==2020.1.16

# System integration
pystray==0.19.4
plyer==2.1.0
PyYAML==6.0.1
python-dateutil==2.8.2

# Async and concurrency
asyncio==3.4.3
aiohttp==3.9.1

# Database
sqlite3  # Built-in
sqlalchemy==2.0.23

# Logging and monitoring
structlog==23.2.0
colorama==0.4.6

# Configuration and CLI
click==8.1.7
pydantic==2.5.2
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.8.0
